requirements.txt
ffmpeg-python
opencv-python
face_alignment
scikit-learn
pydub
pynormalize
soundfile
librosa==0.8.1
pysptk
pyworld
resemblyzer
streamlit
streamlit-webrtc
streamlit-server-state
streamlit-image-select
torchvision
tensorboardX
matplotlib
mediapipe
